python-service/requirements.txt
soccerdata>=1.5.1
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.2.0
python-dotenv>=1.0.0
